requirements.txt
asgiref==3.8.1
django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
